+++ lib/mixins/automatic_random_movement.dart
@@ -9,6 +9,7 @@ mixin AutomaticRandomMovement on Movement {
   static const _KEY_INTERVAL_KEEP_STOPPED = 'INTERVAL_RANDOM_MOVEMENT';
 
   late Random _random;
+  late Random _random2;
 
   /// Method that bo used in [update] method.
   void runRandomMovement(
@@ -28,15 +29,17 @@ mixin AutomaticRandomMovement on Movement {
 
     if (_targetRandomMovement == Vector2.zero()) {
       if (checkInterval(_KEY_INTERVAL_KEEP_STOPPED, timeKeepStopped, dt)) {
-        int randomX = _random.nextInt(maxDistance);
+        int randomX = _random2.nextInt(maxDistance);
         randomX = randomX < minDistance ? minDistance : randomX;
         int randomY = _random.nextInt(maxDistance);
         randomY = randomY < minDistance ? minDistance : randomY;
         final rect = rectConsideringCollision;
         double margin = max(rect.width, rect.height) / 2;
+        int randomNegativeX = _random.nextBool() ? -1 : 1;
+        int randomNegativeY = _random2.nextBool() ? -1 : 1;
         _targetRandomMovement = rect.center.toVector2().translate(
-              (randomX.toDouble() + margin) * _randomNagative(),
-              (randomY.toDouble() + margin) * _randomNagative(),
+              (randomX.toDouble() + margin) * randomNegativeX,
+              (randomY.toDouble() + margin) * randomNegativeY,
             );
         if (useAngle) {
           angle = BonfireUtil.angleBetweenPoints(
@@ -106,11 +109,7 @@ mixin AutomaticRandomMovement on Movement {
   @override
   void onMount() {
     _random = Random(Random().nextInt(1000));
+    _random2 = Random(Random().nextInt(1000));
     super.onMount();
-  }
-
-  int _randomNagative() {
-    int radomInt = _random.nextInt(20);
-    return (radomInt % 2 == 0) ? 1 : -1;
   }
 }